python-app/requirements.txt
fastapi
uvicorn
pydub
speechrecognition
langchain-openai
langchain-core
python-dotenv
python-magic
noisereduce
numpy
tiktoken